Move Controller and Form to new ShareIntend classes
